refactor: Station 필드 자료형 변경
