--- src/main/java/subway/domain/core/Station.java
@@ -1,52 +1,0 @@
-package subway.domain.core;
-
-import java.util.Objects;
-
-public class Station {
-	private Long id;
-	private String name;
-
-	public Station() {
-	}
-
-	public Station(Long id, String name) {
-		this.id = id;
-		this.name = name;
-	}
-
-	public Station(String name) {
-		this.name = name;
-	}
-
-	public Long getId() {
-		return id;
-	}
-
-	public String getName() {
-		return name;
-	}
-
-	@Override
-	public boolean equals(final Object o) {
-		if (this == o)
-			return true;
-		if (o == null || getClass() != o.getClass())
-			return false;
-
-		Station station = (Station)o;
-
-		return Objects.equals(name, station.name);
-	}
-
-	@Override
-	public int hashCode() {
-		return name != null ? name.hashCode() : 0;
-	}
-
-	@Override
-	public String toString() {
-		return "Station{" +
-			"name='" + name + '\'' +
-			'}';
-	}
-}
